25 Oct 2023 Initial comit
ChatGPT3.5 generated code, with manual editions, to plot the "loss" (error) between a line defined as y = wx + b and a cloud of points generated from a linear equation with noise. 

The mean squared error is plotted as a function of the fitting parameters w and b. A machine learning algorithm would find the minimum of the error ("loss" in AI parlance) of in the space of fitting parameters.